Trivially update commit payout image path FREEBIE
Bithub is now serving the payout image with Cache-Control: max-age=0 but
github/fastly is still serving the old cached image. My theory is that
it is not re-reading the headers correctly, although it may be a bug
with github's fastly configuration, since they probably don't expect
anyone to be serving dynamic content in their READMEs.

This commit simply adds a trailing / to the image path, which
effectively forces fastly to re-fetch the image. Hopefully this time
it will read (and respect) the Cache-Control header.