link to actual install instead of blog post
i had all the trouble in the world finding an actual instance of the service running somewhere. only through #19 did i realize there was actually a page listing the projects, how to donate an so on: <https://bithub.herokuapp.com>.

link to it instead of the website, which is linked earlier.